feat: add stage preview screen
